+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 * Disabled blue highlight when Touch/Press `flexberry-objectlistview` or `flexberry-simpleolv`.
+* Fix several words search in `flexberry-lookup` modals.
 
 ## [0.8.0-beta.15] - 2016-12-08
 ### Fixed